Move sequence utilities into ast/seq
