+++ src/CMakeLists.txt
@@ -19,6 +19,7 @@ add_subdirectory(math/subpaving)
 add_subdirectory(ast)
 add_subdirectory(params)
 add_subdirectory(ast/rewriter)
+add_subdirectory(ast/seq)
 add_subdirectory(ast/rewriter/bit_blaster)
 add_subdirectory(ast/normal_forms)
 add_subdirectory(ast/macros)
