+++ src/ast/seq/seq_parikh.h
