+++ src/ast/seq/seq_skolem.h
