+++ src/smt/CMakeLists.txt
@@ -77,6 +77,7 @@ z3_add_component(smt
     uses_theory.cpp
     watch_list.cpp
   COMPONENT_DEPENDENCIES
+    seq
     solver_assertions
     bit_blaster
     cmd_context
