+++ src/smt/seq_axioms.h
@@ -22,8 +22,8 @@ Revision History:
 #include "ast/seq_decl_plugin.h"
 #include "ast/arith_decl_plugin.h"
 #include "ast/rewriter/th_rewriter.h"
-#include "ast/rewriter/seq_skolem.h"
-#include "ast/rewriter/seq_axioms.h"
+#include "ast/seq/seq_skolem.h"
+#include "ast/seq/seq_axioms.h"
 #include "smt/smt_theory.h"
 
 namespace smt {
@@ -108,4 +108,3 @@ namespace smt {
     };
 
 }
-
