+++ src/smt/theory_seq.h
@@ -21,9 +21,9 @@ Revision History:
 #include <optional>
 #include "ast/seq_decl_plugin.h"
 #include "ast/rewriter/th_rewriter.h"
-#include "ast/rewriter/seq_skolem.h"
-#include "ast/rewriter/seq_parikh.h"
-#include "ast/rewriter/seq_eq_solver.h"
+#include "ast/seq/seq_skolem.h"
+#include "ast/seq/seq_parikh.h"
+#include "ast/seq/seq_eq_solver.h"
 #include "ast/ast_trail.h"
 #include "util/scoped_vector.h"
 #include "util/scoped_ptr_vector.h"
@@ -662,4 +662,3 @@ namespace smt {
         bool  get_length(expr* e, rational& r) override;
     };
 }
-
