+++ src/test/seq_eq_approx.cpp
@@ -21,8 +21,8 @@ Author:
 #include "ast/seq_decl_plugin.h"
 #include "ast/arith_decl_plugin.h"
 #include "ast/rewriter/seq_rewriter.h"
-#include "ast/rewriter/seq_eq_approx.h"
-#include "ast/rewriter/seq_view.h"
+#include "ast/seq/seq_eq_approx.h"
+#include "ast/seq/seq_view.h"
 #include "ast/rewriter/seq_regex_witness.h"
 #include "ast/rewriter/expr_safe_replace.h"
 #include "ast/rewriter/th_rewriter.h"
